requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
schedule==1.2.1
pillow
PyQt5==5.15.9
numpy<2.0.0
pyperclip==1.8.2